cleaner refactor, pulling out job handling from API def'n
