+++ backend/app/api.py
@@ -1,15 +1,11 @@
+"""FastAPI interface for a pollable job queue around the generative model."""
 from fastapi import FastAPI
 from fastapi.exceptions import HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse, Response
-import modal
-from modal import Dict
 
-from .common import stub
-from .common import RESULTS_DIR, results_volume
-from .common import get_status, set_status, path_from_job_id
 from .datamodel import JobStatus, JobRequest, JobStatusResponse, HealthResponse
-from .generator import Model
+from . import jobs
 
 
 def create(info) -> FastAPI:
@@ -70,7 +66,8 @@ def create(info) -> FastAPI:
         )
 
         job_id = "-".join([verb, adjective, noun])
-        handle_job(job_id, request)
+
+        jobs.start(job_id, request)
 
         return JobStatusResponse(job_id=job_id, status=JobStatus.PENDING)
 
@@ -83,7 +80,7 @@ def create(info) -> FastAPI:
         tags=["jobs"],
     )
     async def check_job(job_id: str):
-        status = get_status(job_id)
+        status = jobs.check(job_id)
         return JobStatusResponse(job_id=job_id, status=status)
 
     @api_backend.delete(
@@ -94,9 +91,9 @@ def create(info) -> FastAPI:
     )
     async def cancel_job(job_id: str):
         try:
-            set_status(job_id, JobStatus.CANCELLED)
+            jobs.cancel(job_id)
         except KeyError:
-            raise HTTPException(status_code=404, detail="Job not found")
+            raise HTTPException(status_code=404, detail=f"Job {job_id} not found")
         return Response(status_code=204)
 
     @api_backend.get(
@@ -107,65 +104,8 @@ def create(info) -> FastAPI:
     )
     async def read_job(job_id) -> FileResponse:
         try:
-            path = path_from_job_id(job_id)
-            if not path.exists():
-                raise FileNotFoundError
-            return FileResponse(path)
+            return FileResponse(jobs.read(job_id))
         except FileNotFoundError:
             return HTTPException(status_code=404, detail="Job result not found")
 
     return api_backend
-
-
-def save_qr_code(image, path):
-    print(path)
-    return image.save(path)
-
-
-if modal.is_local:
-    stub.jobs = Dict({"_test": {"status": JobStatus.COMPLETE}})
-
-
-@stub.function(timeout=150, shared_volumes={RESULTS_DIR: results_volume}, keep_warm=1)
-def generate_and_save(job_id: str, prompt: str, image: str):
-    """Generate a QR code from a prompt and save it to a file."""
-    try:
-        path = path_from_job_id(job_id)
-        dir_path = path.parent
-        dir_path.mkdir(parents=True, exist_ok=True)
-        call = Model().generate.spawn(prompt, image)
-        set_status(job_id, JobStatus.RUNNING)
-    except Exception as e:
-        set_status(job_id, JobStatus.FAILED)
-        stub.app.jobs[job_id]["error"] = e
-        return
-
-    # await the result
-    try:
-        generator_response = call.get(timeout=60)
-    except TimeoutError as e:
-        set_status(job_id, JobStatus.FAILED)
-        stub.app.jobs[job_id]["error"] = e
-        return
-
-    # write the result to a file
-    save_qr_code(generator_response, path)
-
-    set_status(job_id, JobStatus.COMPLETE)
-
-
-def handle_job(job_id: str, request: JobRequest):
-    try:
-        if job_id == "_test":
-            return
-        else:
-            stub.app.jobs.put(job_id, {"status": JobStatus.PENDING, "handle": None})
-            call = generate_and_save.spawn(
-                job_id, request.prompt, request.image.image_data
-            )
-            stub.app.jobs.put(job_id, {"status": JobStatus.PENDING, "handle": call})
-
-    except Exception as e:
-        print(e)
-        print(f"setting status of {job_id} to failed")
-        set_status(job_id, JobStatus.FAILED)
