+++ backend/app/main.py
@@ -3,8 +3,8 @@ import toml
 
 from .api import create as create_api
 from .common import RESULTS_DIR, results_volume
-from .common import toml_file_path
 from .common import stub
+from .common import toml_file_path
 
 
 @stub.function(
